requirements.txt
pytest==8.4.1
requests==2.32.3
pydantic==2.11.7
allure-pytest==2.14.3
pyAesCrypt==6.1.1
pytest-xdist==3.6.1
allure-pytest==2.14.3
python-dotenv==1.2.1
pytest-html==4.1.1
